% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_launch.R
\name{ca_launch}
\alias{ca_launch}
\title{Launch a caladaptr demo app}
\usage{
ca_launch(
  app = c("timeseries", "chill", "chill2")[1],
  display.mode = c("normal", "showcase")[1]
)
}
\arguments{
\item{app}{App name}

\item{display.mode}{How to display the Shiny app}
}
\description{
Launch a caladaptr demo Shiny app
}
\details{

}
